// src/app/core/enums/deposit-and-withdrawal-types.enum.ts
export enum DepositAndWithdrawalTypes {
  artificialDeposit = 'artificialDeposit', // 人工存款
  otherDeposit = 'otherDeposit', // 其他存款
  cancelPaymentout = 'cancelPaymentout', // 取消出款
  preferentialSupplement = 'Preferential Supplement', // 优惠补充
  clearNegative = 'clearNegative', // 负数清零
  otherPlatformPay = 'otherPlatformPay', // 三方充值
  artificialPaymentout = 'artificialPaymentout', // 人工出款
  otherPaymentout = 'otherPaymentout', // 其他出款
  cancelDeposit = 'cancelDeposit', // 取消入款
  cancelDiscount = 'cancelDiscount', // 取消优惠
  recharge = 'recharge', // 充值
  withdraw = 'withdraw'// 提现
}